+++ src/compiler/typescript_to_rust/azle_generate/src/canister_method/module_item.rs
@@ -1,15 +1,11 @@
 use swc_ecma_ast::{Expr, ExprStmt, FnDecl, ModuleItem};
 
-use crate::{
-    canister_method_annotation::{CanisterMethodAnnotation, CANISTER_METHOD_ANNOTATIONS},
-    ts_ast::GetName,
-};
+use crate::{canister_method::annotation::CANISTER_METHOD_ANNOTATIONS, ts_ast::GetName};
 
 pub trait ModuleItemHelperMethods {
     fn is_custom_decorator(&self) -> bool;
     fn as_exported_fn_decl(&self) -> Option<FnDecl>;
     fn as_expr_stmt(&self) -> Option<ExprStmt>;
-    fn to_canister_method_annotation(&self) -> Option<CanisterMethodAnnotation>;
 }
 
 impl ModuleItemHelperMethods for ModuleItem {
@@ -59,16 +55,5 @@ impl ModuleItemHelperMethods for ModuleItem {
 
     fn as_expr_stmt(&self) -> Option<ExprStmt> {
         Some(self.as_stmt()?.as_expr()?.clone())
-    }
-
-    fn to_canister_method_annotation(&self) -> Option<CanisterMethodAnnotation> {
-        let method_type = self
-            .as_stmt()?
-            .as_expr()?
-            .expr
-            .as_ref()
-            .as_ident()?
-            .get_name();
-        CanisterMethodAnnotation::new(method_type, None).ok()
     }
 }
