+++ src/compiler/typescript_to_rust/azle_generate/src/ts_ast/mod.rs
@@ -14,7 +14,6 @@ mod azle_functions_and_methods;
 mod azle_new_expr;
 mod expr;
 mod ident;
-mod module;
 mod ts_ast;
 mod ts_fn_param;
 mod ts_fn_type;
@@ -27,7 +26,6 @@ pub mod azle_type;
 pub mod azle_type_alias_decls;
 pub mod class_decl;
 pub mod class_prop;
-pub mod module_item;
 pub mod param;
 pub mod source_map;
 pub mod traits;
